moved file writing functionality into new header "WriteFile". originally attempted to add the function to the "FileHelpers" header, however, the fact that the "PpmDocument" class depends on it led to errors when I tried to include "PpmDocument" in "FileHelpers". as such, I opted to just create a new header. I might rename "FileHelpers" to "ReadFile" for program legibility / clarity in the next commit.

+++ PpmDocument.hpp
@@ -1,12 +1,13 @@
 #pragma once
 #ifndef PPM_DOCUMENT_HPP
 #define PPM_DOCUMENT_HPP
+
 #include <string>
 #include <vector>
 #include <sstream>
 #include <fstream>
-#include "FileHelpers.h"
 #include "Pixel.hpp"
+#include "FileHelpers.h"
 using namespace std;
 
 class PpmDocument
@@ -133,7 +134,7 @@ public:
 
 	int getPixelCount() const
 	{
-		return _pixel_data.size();
+		return static_cast<int>(_pixel_data.size());
 	}
 
 };
